+++ source/TriangleData.h
@@ -9,8 +9,8 @@ public:
     static void GenVertexData();
 
     static std::vector<Vertex>& GetVertexData();
-    static std::vector<uint16_t>& GetIndexData();
+    static std::vector<uint32_t>& GetIndexData();
 private:
     inline static std::vector<Vertex> vertices;
-    inline static std::vector<uint16_t> indices;
+    inline static std::vector<uint32_t> indices;
 };
